I integrated Font Awesome 3.0 into the website.
I added Font Awesome as a Git submodule to the website project. I
updated the Rakefile script to copy the Font Awesome font files into
the /css/fonts directory. I imported the Font Awesome LESS file in the
normal theme and corrected the path to the Font Awesome font files in
the /css/fonts directory.