I fixed a rendering problem due to Reveal.js.
There was a rendering problem because of conflicting CSS with the
Reveal.js CSS stylesheet. I removed the CSS for reveal.js from the
normal theme and modified the Rakefile script to copy reveal.min.css
into the css directory. I changed the presentation template to include
the reveal.min.css stylesheet.