Refactor website
I implemented a major refactoring and modernization of the website.

I removed all of the old Git submodules.

I replaced the build system to use Grunt instead of Rake.

I'm managing dependencies using NPM and Bower.

I changed the layout of the homepage, the blog index pages, and the
post page.