Fix website responsiveness
I modified the CSS/LESS rules for the normal theme to make the website
more "responsive" to different device sizes. I modified the jumbotron
on the homepage so that the background cover photo only appears for
tablets and larger screens. On mobile phones, the background will not
appear and the text in the jumbotron will not be pushed down.