Simplify blog posts
I simplified the blog posts. I removed the description front matter
variable and replaced it with the post's excerpt variable in the head
metadata for the post.

I removed the author metadata information used for Twitter and
OpenGraph. I set defaults for the metadata in the _config.yml file
so that it does not need to be repeated for every post.