Add recent posts to home page
I modified the homepage for the website. I split the content area into
two columns. The text of the latest post is in the left column. A list
of the most recent next 5 posts is in the right column.

I modified the Stuck in a Rut post. I reduced the size of the embedded
TED video to render better in the homepage. The larger video was
flowing into the right column and covering the recent posts.