New blog post: Refactoring My Go Web Server
I wrote a new blog post about refactoring my Go web API server to make
it better and easier to maintain.

I fixed a problem with the post template for older posts without a
cover photo. The previous and next buttons at the top of the blog post
were covered by the fixed navigation bar. I modified the style sheet
and the post template to make the navigation buttons visible.