Fix excerpt in post layout
I fixed an error in the metadata for the post template. The excerpt
that was being generated by Jekyll included the HTML p tag at the
beginning and end of the excerpt. I modified the metadata to strip the
begin and end tag.

I fixed the description meta element in the header. It was not
using the new excerpt property on the page object, causing the excerpt
to not be output in the generated page output.