Configure jekyll-feed Plugin
I added support for the jekyll-feed plugin to generate the ATOM feed for
the website. I added site and author configuration settings to
_config.yml to customize the feed output.

I updated _config.yml. I changed the brand variable to title.

I restored the RSS feed layout and post include in order to get the
legacy RSS feed to generate.